requirements.txt
requests==2.22.0
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision
Pillow>=5.2
flask